backend/requirements.txt
# BookInsight backend — install with: pip install -r requirements.txt
# (Use a virtual environment; first install may download PyTorch via sentence-transformers.)

Django>=4.2,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
selenium>=4.15,<5
requests>=2.31,<3
chromadb>=0.4.24,<1
sentence-transformers>=2.2,<6
